+++ internal/manager/controllerdef.go
@@ -292,6 +292,7 @@ func setupControllers(
 				DisableIngressClassLookups: !c.IngressClassNetV1Enabled,
 				CacheSyncTimeout:           c.CacheSyncTimeout,
 				ReferenceIndexers:          referenceIndexers,
+				StatusQueue:                kubernetesStatusQueue,
 			},
 		},
 		{